Fixes based on merge request feedback

+++ include/gpstolfp.h
@@ -7,5 +7,5 @@
 
 extern void gpstolfp (int weeks, int days, unsigned long seconds, l_fp *);
 extern void gpsweekadj (unsigned int * week, unsigned int build_week);
-extern void gpstocal (unsigned int week, unsigned int TOW, int UTC_offset, struct calendar *);
-extern void caltogps (const struct calendar *, int UTC_offset, unsigned int * week, unsigned int * TOW);
+extern void gpstocal (unsigned int week, unsigned long int TOW, int UTC_offset, struct calendar *);
+extern void caltogps (const struct calendar *, int UTC_offset, unsigned int * week, unsigned long int * TOW);
